+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.10.0] - 2026-09-13
+
 ### Changed
 
 - **Info screen: collapsed the api/app version badges into one.** Since the